+++ 1-foundation/practical-logger-core/src/consts.ts
@@ -29,34 +29,17 @@ export const LOG_LEVEL_TO_INTEGER: Readonly<{ [k: string]: number }> = {
 	silly:  100,
 }
 
-// rationalization to a clear, human understandable string
-export const LOG_LEVEL_TO_HUMAN: Readonly<{ [k: string]: string }> = {
-	fatal:   'fatal',
-	emerg:   'emergency',
-
-	alert:   'alert',
-	crit:    'critical',
-	error:   'error',
-
-	warning: 'warn', // warn is a word, shorter is better
-	warn:    'warn',
-
-	notice:  'notice',
-
-	info:    'info',
-
-	verbose: 'verbose',
-	log:     'log',
-	debug:   'debug',
-
-	trace:   'trace',
-
-	silly:   'silly',
-}
-
 export const ALL_LOG_LEVELS: Readonly<LogLevel[]> =
 	Object.keys(LOG_LEVEL_TO_INTEGER)
 		.map(s => s as LogLevel)
 		.sort((a: LogLevel, b: LogLevel) => LOG_LEVEL_TO_INTEGER[a] - LOG_LEVEL_TO_INTEGER[b])
+
+// rationalization to a clear, human understandable string
+export const LOG_LEVEL_TO_HUMAN: Readonly<{ [k: string]: string }> =
+	// generated to shave a few bytes
+	Object.fromEntries(
+		ALL_LOG_LEVELS
+			.map((ll) => { return [ ll, ({ em: 'emergency', wa: 'warn'} as any)[ll.slice(0, 1)] || ll ]})
+	);
 
 export * from './consts-base'