+++ 1-foundation/practical-logger-core/src/normalize-args.ts
@@ -4,93 +4,54 @@ import {
 
 
 export function looksLikeAnError(x: any): boolean {
-	return !!(x.name && x.message && x.stack)
+	return !!(x && x.name && x.message && x.stack)
 }
 
 // harmonize
-// also try to recover from some common errors
-// TODO assess whether it's really good to be that permissive (also: hurts perfs)
-export function normalizeArguments(args: IArguments): [ string, LogDetails ] {
-	//console.log('>>> NA', Array.from(args))
-
-	let message: string = ''
-	let details: Readonly<LogDetails> = {}
+// also try to recover from incorrect invocations
+export function normalizeArguments(raw_args: IArguments): [ string, LogDetails ] {
+	let message_parts: string[] = []
+	let details: LogDetails = {}
 	let err: Error | undefined = undefined
 
-	if (args.length > 2) {
-		//console.warn('NA 1', args)
-		// wrong invocation,
-		// most likely a "console.log" style invocation from an untyped codebase.
-		// "best effort" fallback:
-		message = Array.prototype.join.call(args, ' ')
-		details = {}
+	Array.from(raw_args)
+		.forEach(arg => {
+			if (!arg)
+				return
+
+			// errors are first class, look for them first
+			if (looksLikeAnError(arg)) {
+				if (!err)
+					err = arg // extract it
+				return
+			}
+			if (!err && looksLikeAnError(arg.err)) {
+				err = arg.err // extract it
+				// don't return, still stuff to pick
+			}
+
+			if (typeof arg === 'object') {
+				details = {
+					...details,
+					...arg,
+				}
+
+				return
+			}
+
+			message_parts.push(String(arg))
+		})
+
+	if (typeof details.message === 'string' && !message_parts.length) {
+		message_parts.push(details.message)
+		delete details.message
 	}
-	else {
-		//console.log('NA 2')
-		message = args[0] || ''
-		details = args[1] || {}
 
-		// optimization
-		if (!message || typeof args[0] !== 'string' || typeof details !== 'object') {
-			// non-nominal call
-			//console.warn('NA 2.1')
-
-			// try to fix message (attempt 1)
-			if (typeof message !== 'string') {
-				//console.warn('NA 2.1.1', { message, details })
-				if (looksLikeAnError(message)) {
-					//console.warn('NA 2.1.1.1')
-					// Another bad invocation
-					// "best effort" fallback:
-					err = message as Error
-					message = err.message
-				}
-				else if (typeof message === 'object' && !args[1]) {
-					// no message, direct details
-					//console.warn('NA 2.1.1.2')
-					details = message as LogDetails
-					message = ''
-				}
-				else {
-					//console.warn('NA 2.1.1.3')
-					message = String(message)
-				}
-			}
-
-			// try to fix details
-			if (typeof details !== 'object') {
-				//console.warn('NA 2.1.2', { details })
-				// Another bad invocation
-				// "best effort" fallback:
-				message = [ message, String(details) ].join(' ')
-				details = {}
-			}
-
-			// ensure we picked up err
-			err = err || details.err
-
-			// attempt to fix message (attempt 2, after uniformizing details)
-			if (!message && details.message) {
-				//console.warn('NA 2.1.3', { details })
-				const { message: m2, ...d2 } = details
-				message = m2
-				details = d2
-			}
-
-			message = message || (err && err.message) || '(no message)'
-		}
-
-		if (!err && looksLikeAnError(details)) {
-			//console.warn('NA 2.2', { details })
-			// details is in fact an error, extract it
-			err = details as Error
-			details = { err }
-		}
-		else if (err)
-			details = { err, ...details }
-		else
-			details = { ...details }
-	}
+	const message = message_parts.join(' ') || (err as any)?.message || '(no message)'
+	if (err)
+		details.err = err
+	else
+		delete details.err // because could be present but not be a correct err type
 
 	return [ message, details ]
 }